Implement constant time modular addition
This requires the use of a constant time copy routine, which is adapated
from go's subtle version, which works for bytes, in order to work with
Word instead.

One thing I realized was that we don't actually assume that x and y are
with the range [0, m). This assumption would speed things up, since we
can avoid a round of modular reduction for x and y.

It might be a good idea to allow this faster mode of operation.
Two avenues open up here:
a) Make this the behavior for ModAdd
  - Pros: Avoid creating an extra method
  - Cons: error prone usage
b) Add a new method
  - Pros: Obvious which one to use
  - Cons: still possible to make mistakes, now you have two methods