Implement Modular Exponentiation
This is a standard bit by bit method, and should be satisfactory,
provided ModMul is fast, which isn't exactly the case now

This revealed a bug in our implementation of limb resizing, which would
accidentally modify the length of a supposedly immutable nat. The fix
was to create a slice with a longer capacity, but the same length.