Use a different type for Moduli
This type is also allowed to leak the exact bit length.

In most cases, the modulus is publicly known, so this should be fine.
In the cases where the value is secret, for example, when using
the totient as modulus in RSA, the exact bit length is still leakable,
because the approximate value is still well known.

Regardless, having a different Modulus type is probably a good idea,
in order to give ourselves the evolutionary freedom to do things like
this.

Anyways, the reasoning behind this change is twofold. The first is that
since this assumption, of being able to leak the true length, is
reasonable, it also speeds up any operation with modular arithmetic,
since we don't waste space and thus time by having an exactly sized
modulus.

Furthermore, it seems that limb by limb modular reduction requires
knowing the exact bit length, and having a "tightly sized" modulus. This
is much desirable over the current bit by bit reduction.

I think, in theory, you could probably orchestrate a constant time
version of this, but this would involve constant time table lookups over
the entire accumulator when doing reduction. I think. This isn't all
that desirable.

That being said, if this security assumption is indeed not reasonable,
then performance doesn't matter, as compared to security. It seems that
is is a reasonable assumption, and BearSSL does it this way, notably.