Implement less leaky leadingZeros method
This should be less leaky because we avoid comparing the limbs with
values in a non-CT way. We can't avoid leaking the number of leading
limbs, but that is part of the contract, anyways.

This kind of leakage should be fine in theory, since the main use cases
for this function are validation.